Menambahkan beberapa fitur untuk halaman admin

+++ database/migrations/2025_09_12_025524_add_foto_to_users_table.php
@@ -9,14 +9,18 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->string('foto')->nullable()->after('password'); // kolom foto, nullable
+            if (!Schema::hasColumn('users', 'foto')) {
+                $table->string('foto')->nullable()->after('email');
+            }
         });
     }
 
     public function down(): void
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('foto');
+            if (Schema::hasColumn('users', 'foto')) {
+                $table->dropColumn('foto');
+            }
         });
     }
 };
